requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
scikit-learn==1.5.1
scipy==1.14.1
umap-learn==0.5.6
